refactor: update BookData Type
