+++ apps/web/src/routes/kiwi/-reader/contexts/settings-context.tsx
@@ -12,7 +12,7 @@ import { updateCustomStyle } from "../styles";
 
 import { useBook } from "./book-context";
 
-import { Settings } from "#/types/reader";
+import { Settings } from "#/types/book";
 
 interface SettingsContextType extends Settings {
   setIsSinglePage: (isSinglePage: boolean) => void;
